fix random, and update latex

+++ disJ/src/distributed/plugin/random/IRandom.java
@@ -2,7 +2,7 @@ package distributed.plugin.random;
 
 public interface IRandom {
 	
-	public int nextInt( double n );
-	public void setSeed( long seed );
+	public int nextInt(int n);
+	public void setSeed(long seed);
 	
 }
